requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-redis==5.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
mypy==1.8.0
django-stubs==4.2.7
djangorestframework-stubs==3.14.5
types-redis==4.6.0.20240311
types-requests==2.31.0.20240311
types-python-dateutil==2.8.19.20240311
drf-yasg==1.21.7
pyopenssl==23.2.0
django-extensions==3.2.3
werkzeug==2.3.7

# Deployment dependencies
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0

# Security and monitoring dependencies
sentry-sdk==1.39.1
django-prometheus-metrics==2.2.1
python-json-logger==2.0.7
django-axes==6.1.1
safety==2.3.5
django-csp==3.7